+++ sqlitex/query.go
@@ -9,8 +9,10 @@ import (
 	"zombiezen.com/go/sqlite"
 )
 
-var errNoResults = errors.New("sqlite: statement has no results")
-var errMultipleResults = errors.New("sqlite: statement has multiple result rows")
+var (
+	errNoResults       = errors.New("sqlite: statement has no results")
+	errMultipleResults = errors.New("sqlite: statement has multiple result rows")
+)
 
 func resultSetup(stmt *sqlite.Stmt) error {
 	hasRow, err := stmt.Step()
@@ -99,4 +101,19 @@ func ResultFloat(stmt *sqlite.Stmt) (float64, error) {
 		return 0, err
 	}
 	return res, nil
+}
+
+// ResultBytes reads the first column of the first and only row
+// produced by running stmt into buf,
+// returning the number of bytes read.
+// It returns an error if there is not exactly one result row.
+func ResultBytes(stmt *sqlite.Stmt, buf []byte) (int, error) {
+	if err := resultSetup(stmt); err != nil {
+		return 0, err
+	}
+	read := stmt.ColumnBytes(0, buf)
+	if err := resultTeardown(stmt); err != nil {
+		return 0, err
+	}
+	return read, nil
 }